Bridges for Tyruswoo_TileControl.js MZ
Bridges added to Tile Control MZ.  This feature is working fairly well, but deserves more refinement before release.

The user must define when tile tag indicates bridges, and if no terrain tag is given for bridges, all of these bridge features should be off.  So, if there is no terrain tag for bridges, this code results in identical functionality as Tile Control had previously without the extra calculations of bridge passability.

Note that bridges result in multiple altitudes existing on a map, so that events and player characters can be above or below each other.